requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosmtplib==3.0.2
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
assemblyai==0.32.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
docx==0.2.4
exceptiongroup==1.2.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
lxml==5.3.0
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
openai==0.28.0
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyTelegramBotAPI==4.21.0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
regex==2024.5.15
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
telebot==0.0.5
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
websockets==13.0
yarl==1.9.4
